# openapi.yaml
openapi: 3.0.1
info:
  title: Catatan Keuangan Plugin
  description: Plugin untuk mencatat pengeluaran dan pemasukan
  version: "1.0.0"
servers:
  - url: https://7071-101-255-150-125.ngrok-free.app

paths:
  /login:
    post:
      summary: Login untuk mendapatkan token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Token berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Gagal login

  /transactions:
    get:
      summary: Lihat semua transaksi
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Daftar transaksi
    post:
      summary: Tambah transaksi baru
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [income, expense]
                amount:
                  type: number
                description:
                  type: string
      responses:
        "201":
          description: Transaksi berhasil ditambahkan

  /balance:
    get:
      summary: Lihat total pemasukan, pengeluaran, dan saldo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Ringkasan saldo

  /export/pdf:
    get:
      summary: Unduh laporan transaksi dalam bentuk PDF
      security:
        - bearerAuth: []
      responses:
        "200":
          description: File PDF berisi laporan transaksi

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT